requirements.txt
Django==4.1.3
django-cors-headers==3.13.0
django-extensions==3.0.2
django-filter==22.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
